+++ lumicks/pylake/tests/data/mock_widefield.py
@@ -46,8 +46,10 @@ class MockTiffFile:
         return len(self._src.pages)
 
 
-def make_frame_times(n_frames, step=8, start=10):
-    return [[f"{j}", f"{j+step}"] for j in range(start, start + (n_frames + 1) * 100, 10)]
+def make_frame_times(n_frames, step=8, start=10, frame_time=10):
+    return [
+        [f"{j}", f"{j+step}"] for j in range(start, start + (n_frames + 1) * frame_time, frame_time)
+    ]
 
 
 def apply_transform(spots, Tx, Ty, theta, offsets=None):